feat: filter items with select display option
- filter items
- add confirm pop up for delete event
- fix little bug (remainDays at server)

+++ src/pages/ItemForm.tsx
@@ -110,6 +110,10 @@ const ItemForm = ({ apiURL }: Props) => {
 
   // function handle form delete
   const handleDelete = async () => {
+    const isConfirmed = confirm("Are you sure you want to delete this item?");
+    if (!isConfirmed) {
+      return;
+    }
     try {
       await axios.delete(`${apiURL}/item/${itemId}`);
       alert("Item delete successfully");
